chg: Move io functions to io module
